Simulate Data from Multivariate Normal Means Model

Simulate data points from the multivariate normal means model. The univariate case (m = 1) is also handled. See ud_fit for the model definition.

simulate_ud_data(n, w, U, S)

Arguments

n
Number of data points to simulate.

w
A numeric vector of length k specifying the prior mixture weights. All entries must be non-negative, but need not sum to 1; the mixture weights are automatically normalized to sum to 1.

U
A list of length k specifying the covariance matrices in the mixture-of-multivariate-normals prior; list element U[[i]] is the m x m covariance matrix for the ith mixture component.

S
The m x m residual covariance matrix.

Value

An n x m matrix in which each row is a draw from the multivariate normal means model. For the univariate case (m = 1), a vector is returned.

See also