Fit Ultimate Deconvolution Model

This function implements "Ultimate Deconvolution", an empirical Bayes method for fitting a multivariate normal means model. This method is closely related to approaches for multivariate density deconvolution (Sarkar et al, 2018), so it can also be viewed as a method for multivariate density deconvolution.

ud_fit(fit0, X, control = list(), verbose = TRUE)

ud_fit_control_default()

Arguments

fit0
A previous Ultimate Deconvolution model fit. Typically, this will be an output from ud_init or from a previous call to ud_fit.

X
Describe input argument "X" here.

control
A list of parameters controlling the behaviour of the model fitting and initialization. See ‘Details’.

verbose
When verbose = TRUE, information about the algorithm's progress is printed to the console at each iteration. For interpretation of the columns, see the description of the progress return value.

Value

A list object with the following elements:

w

A vector containing the estimated prior mixture weights. When control$update.w = "none", this will be the same as the w provided at input.

U

A list containing the estimated prior covariance matrices. When control$update.U = "none", this will be the same as the U provided at input.

V

The estimated residual covariance matrix. When control$update.S = "none", this will be the same as the S provided at input.

loglik

The log-likelihood at the current settings of the model parameters, w, U and S.

progress

A data frame containing detailed information about the algorithm's progress. The columns of the data frame are: "iter", the iteration number; "loglik", the log-likelihood at the current estimates of the model parameters; "delta.w", the largest change in the mixture weights; "delta.u", the largest change in the prior covariance matrices; "delta.s", the largest change in the residual covariance matrix; and "timing", the elapsed time in seconds (recorded using proc.time).

Details

In the Ultimate Deconvolution (UD) model, the m-dimensional observation \(x\) is drawn from a mixture of multivariate normals, \(x ~ w_1 N(0, T_1) + ... + w_k N(0, T_k)\), where \(k\) is the number of mixture components, the \(w_j\)'s are the mixture weights, and each \(T_j = V + U_j\) is a covariance matrix. This is the marginal density derived from a model in which \(x\) is multivariate normal with mean \(y\) and covariance \(V\), and the underlying, or "latent", signal \(y\) is in turn modeled by a mixture prior in which each mixture component \(j\) is multivariate normal with zero mean and covariance matrix \(U_j\). This model is a useful special case of the "Extreme Deconvolution" (ED) model (Bovy et al, 2011).

Two variants of the UD model are implemented: one in which the residual covariance V is the same for all data points, and another in which V is different for each data point. In the first case, this covariance matrix may be estimated.

The UD model is fit by expectation-maximization (EM). The control argument is a list in which any of the following named components will override the default optimization algorithm settings (as they are defined by ud_fit_control_default):

weights.update

When weights.update = "em", the mixture weights are updated via EM; when weights.update = "none", the mixture weights are not updated.

version

R and C++ implementations of the model fitting algorithm are provided; these are selected with version = "R" and version = "Rcpp".

maxiter

The upper limit on the number of EM updates to perform.

tol

Convergence tolerance for the EM algorithm; the updates are halted when the largest change in the model parameters between two successive updates is less than tol.

update.U

This setting determines the algorithm used to estimate the prior covariance matrices. Two EM variants are implemented: update.U = "ed", the EM algorithm described by Bovy et al (2011); and update.U = "teem" (truncated eigenvalue expectation-maximization), in which the M-step update for each covariance matrix U[[j]] is solved by truncating the eigenvalues in a spectral decomposition of the unconstrained maximimum likelihood estimate (MLE) of U[j]]. The latter method provides greater freedom in the updates for U, and is expected to yield better fits, and converge more quickly to a good fit.

update.S

When update.S = "em", maximum-likelihood estimate of the residual covariance matrix is computed via EM; when update.S = "none", the residual covariance parameter is not updated.

minval

Minimum eigenvalue allowed in the residual covariance(s) V and the prior covariance matrices U.

Using this function requires some care; currently only minimal argument checking is performed. See the documentation and examples for guidance.

References

J. Bovy, D. W. Hogg and S. T. Roweis (2011). Extreme Deconvolution: inferring complete distribution functions from noisy, heterogeneous and incomplete observations. Annals of Applied Statistics, 5, 1657–1677. doi:10.1214/10-AOAS439

A. Sarkar, D. Pati, A. Chakraborty, B. K. Mallick and R. J. Carroll (2018). Bayesian semiparametric multivariate density deconvolution. Journal of the American Statistical Association 113, 401–416. doi:10.1080/01621459.2016.1260467

J. Won, J. Lim, S. Kim and B. Rajaratnam (2013). Condition-number-regularized covariance estimation. Journal of the Royal Statistical Society, Series B 75, 427–450. doi:10.1111/j.1467-9868.2012.01049.x

Examples

# TO DO.