Fit Empirical Bayes Multivariate Normal Means Model

This function implements "Ultimate Deconvolution", an empirical Bayes method for fitting a multivariate normal means model. This method is closely related to approaches for multivariate density deconvolution (Sarkar et al, 2018), so it can also be viewed as a method for multivariate density deconvolution.

ud_fit(X, k, fit0, w, U, S = diag(ncol(X)), control = list(), verbose = TRUE)

ud_fit_control_default()

Arguments

X
The n x m data matrix, in which each row is an m-dimensional observation.

k
An integer 2 or greater specifying the number of components in the mixture-of-normals prior. This only needs to be provided if neither U nor w are provided.

fit0
A previous Ultimate Deconvolution fit. This is useful for "re-fitting" the model. When a value for this argument is given, the other inputs k, w, U and S are not needed.

w
A numeric vector of length k giving initial estimates of the prior mixture weights. All entries must be non-negative, but need not sum to 1; the mixture weights are automatically normalized to sum to 1. If not provided, the mixture weights are set to uniform.

U
A list of length k giving initial estimates of the covariance matrices in the mixture-of-multivariate-normals prior; list element U[[i]] is the m x m covariance matrix for the ith mixture component. If not provided, the initial estimates are randomly generated by computing sample covariances on random subsets of the data X.

S
An m x m matrix giving the initial estimate of the residual covariance matrix.

control
A list of parameters controlling the behaviour of the model fitting algorithm. See ‘Details’.

verbose
When verbose = TRUE, information about the algorithm's progress is printed to the console at each iteration. For interpretation of the columns, see the description of the progress return value.

Value

A list object with the following elements:

w

A vector containing the estimated prior mixture weights. When control$update.w = "none", this will be the same as the w provided at input.

U

A list containing the estimated prior covariance matrices. When control$update.U = "none", this will be the same as the U provided at input.

S

The estimated residual covariance matrix. When control$update.S = "none", this will be the same as the S provided at input.

loglik

The log-likelihood at the current settings of the model parameters, w, U and S.

progress

A data frame containing detailed information about the algorithm's progress. The columns of the data frame are: "iter", the iteration number; "loglik", the log-likelihood at the current estimates of the model parameters; "delta.w", the largest change in the mixture weights; "delta.u", the largest change in the prior covariance matrices; "delta.s", the largest change in the residual covariance matrix; and "timing", the elapsed time in seconds (recorded using proc.time).

Details

In the multivariate normal means model, each m-dimensional observation \(x\) is drawn from a mixture of multivariate normals, \(x ~ w_1 N(0, T_1) + ... + w_k N(0, T_k)\), where \(k\) is the number of mixture components, the \(w_j\)'s are the mixture weights, and each \(T_j = S + U_j\) is a covariance matrix. This is the marginal density derived from a model in which each \(x\) is multivariate normal with mean \(y\) and covariance \(S\), and the underlying, or "latent", signal \(y\) is in turn modeled by a mixture prior in which each mixture component \(j\) is multivariate normal with zero mean and covariance matrix \(S_j\). The "Extreme Deconvolution" (ED) model (Bovy et al, 2011) is a slight generalization of this multivariate normal means model; the ED model allows for the mixture components to have nonzero means, sample-specific residual covariances, and it allows one to specify a linear projection of the underlying signal onto the observed signal. Therefore, this method also implements a useful special case of Extreme Deconvolution.

The multivariate normal means model is fit by expectation-maximization (EM). The control argument is a list in which any of the following named components will override the default optimization algorithm settings (as they are defined by ud_fit_control_default):

update.w

When update.w = "em", maximum-likelihood estimates of the mixture weights are computed via an EM algorithm; when update.w = "mixsqp", the mix-SQP solver is used to update the mixture weights by maximizing the likelihood with the other parameters fixed; when update.w = "none", the mixture weights are not updated.

update.U

This setting determines the algorithm used to estimate the prior covariance matrices. Two EM variants are implemented: update.U = "ed", the EM algorithm described by Bovy et al (2011); and update.U = "teem" (truncated eigenvalue expectation-maximization), in which the M-step update for each covariance matrix U[[j]] is solved by truncating the eigenvalues in a spectral decomposition of the unconstrained maximimum likelihood estimate (MLE) of U[j]]. The latter method provides greater freedom in the updates for U, and is expected to yield better fits, and converge more quickly to a good fit.

update.S

When update.S = "em", maximum-likelihood estimate of the residual covariance matrix is computed via EM; when update.S = "none", the residual covariance parameter is not updated.

version

R and C++ implementations of the model fitting algorithm are provided; these are selected with version = "R" and version = "Rcpp".

maxiter

The upper limit on the number of EM updates to perform.

tol

Convergence tolerance for the EM algorithm; the updates are halted when the largest change in the model parameters between two successive iterations of EM is less than tol.

minval

A small, non-negative number specifying the lower bound on the eigenvalues of the prior covariance matrices U.

Using this function requires some care; currently only minimal argument checking is performed. See the documentation and examples for guidance.

References

J. Bovy, D. W. Hogg and S. T. Roweis (2011). Extreme Deconvolution: inferring complete distribution functions from noisy, heterogeneous and incomplete observations. Annals of Applied Statistics, 5, 1657–1677. doi:10.1214/10-AOAS439

A. Sarkar, D. Pati, A. Chakraborty, B. K. Mallick and R. J. Carroll (2018). Bayesian semiparametric multivariate density deconvolution. Journal of the American Statistical Association 113, 401–416. doi:10.1080/01621459.2016.1260467

J. Won, J. Lim, S. Kim and B. Rajaratnam (2013). Condition-number-regularized covariance estimation. Journal of the Royal Statistical Society, Series B 75, 427–450. doi:10.1111/j.1467-9868.2012.01049.x

Examples

library(mvtnorm)
set.seed(1)
X   <- rmvt(1000,diag(2),df = 4)
fit <- ud_fit(X,k = 10,S = diag(2))

#> Performing Ultimate Deconvolution on 1000 x 2 matrix with these settings:
#> number of components in mixture prior: 10max 100 updates, conv tol 1.0e-06 (udr 0.2-3, "Rcpp").
#> updates: w (mix weights) = em; U (prior cov) = teem; S (resid cov) = none
#> iter          log-likelihood |w - w'| |U - U'| |S - S'|
#>    1 -3.4066145888315705e+03 1.38e-02 1.41e+00 0.00e+00
#>    2 -3.3643627844578314e+03 9.05e-03 2.11e+00 0.00e+00
#>    3 -3.3452731145221537e+03 9.25e-03 1.40e+00 0.00e+00
#>    4 -3.3362435394971985e+03 7.84e-03 9.64e-01 0.00e+00
#>    5 -3.3325260294960294e+03 6.50e-03 8.03e-01 0.00e+00
#>    6 -3.3306348867410620e+03 5.47e-03 8.04e-01 0.00e+00
#>    7 -3.3292757882216165e+03 4.68e-03 8.14e-01 0.00e+00
#>    8 -3.3281849744722604e+03 4.03e-03 8.20e-01 0.00e+00
#>    9 -3.3272856134344756e+03 3.47e-03 8.25e-01 0.00e+00
#>   10 -3.3265356322057578e+03 3.00e-03 8.24e-01 0.00e+00
#>   11 -3.3259024639598138e+03 2.72e-03 8.14e-01 0.00e+00
#>   12 -3.3253588194093481e+03 2.46e-03 7.96e-01 0.00e+00
#>   13 -3.3248829258062060e+03 2.22e-03 7.72e-01 0.00e+00
#>   14 -3.3244806354684602e+03 2.01e-03 7.47e-01 0.00e+00
#>   15 -3.3241429776075925e+03 1.83e-03 7.29e-01 0.00e+00
#>   16 -3.3238402322435418e+03 1.67e-03 7.14e-01 0.00e+00
#>   17 -3.3235625552844276e+03 1.52e-03 6.93e-01 0.00e+00
#>   18 -3.3233046288811147e+03 1.39e-03 6.74e-01 0.00e+00
#>   19 -3.3230890782630318e+03 1.28e-03 6.59e-01 0.00e+00
#>   20 -3.3228965235942019e+03 1.19e-03 6.58e-01 0.00e+00
#>   21 -3.3227148936344724e+03 1.10e-03 6.58e-01 0.00e+00
#>   22 -3.3225620733941678e+03 1.03e-03 6.55e-01 0.00e+00
#>   23 -3.3224320434146002e+03 9.69e-04 6.58e-01 0.00e+00
#>   24 -3.3223093740785580e+03 9.18e-04 6.60e-01 0.00e+00
#>   25 -3.3221932124799250e+03 8.76e-04 6.56e-01 0.00e+00
#>   26 -3.3220836952751715e+03 8.41e-04 6.47e-01 0.00e+00
#>   27 -3.3219812661194965e+03 8.13e-04 6.37e-01 0.00e+00
#>   28 -3.3218862704635667e+03 7.92e-04 6.24e-01 0.00e+00
#>   29 -3.3217987088908130e+03 7.78e-04 6.05e-01 0.00e+00
#>   30 -3.3217181460057836e+03 7.68e-04 5.81e-01 0.00e+00
#>   31 -3.3216437458998312e+03 7.63e-04 5.50e-01 0.00e+00
#>   32 -3.3215743811889020e+03 7.61e-04 5.13e-01 0.00e+00
#>   33 -3.3215087615696666e+03 7.62e-04 4.71e-01 0.00e+00
#>   34 -3.3214455443547536e+03 7.65e-04 4.27e-01 0.00e+00
#>   35 -3.3213834109149761e+03 7.71e-04 3.83e-01 0.00e+00
#>   36 -3.3213210790145658e+03 7.77e-04 3.40e-01 0.00e+00
#>   37 -3.3212572190931151e+03 7.86e-04 3.00e-01 0.00e+00
#>   38 -3.3211905552612807e+03 7.98e-04 2.64e-01 0.00e+00
#>   39 -3.3211199769986183e+03 8.13e-04 2.32e-01 0.00e+00
#>   40 -3.3210446005375270e+03 8.30e-04 2.05e-01 0.00e+00
#>   41 -3.3209638466632705e+03 8.48e-04 1.83e-01 0.00e+00
#>   42 -3.3208775217402567e+03 8.68e-04 1.64e-01 0.00e+00
#>   43 -3.3207858777156052e+03 8.89e-04 1.67e-01 0.00e+00
#>   44 -3.3206896216867272e+03 9.09e-04 1.82e-01 0.00e+00
#>   45 -3.3205898532989650e+03 9.27e-04 1.99e-01 0.00e+00
#>   46 -3.3204879307584838e+03 9.42e-04 2.18e-01 0.00e+00
#>   47 -3.3203852952840821e+03 9.52e-04 2.40e-01 0.00e+00
#>   48 -3.3202833027638344e+03 9.56e-04 2.63e-01 0.00e+00
#>   49 -3.3201831071431052e+03 9.53e-04 2.89e-01 0.00e+00
#>   50 -3.3200856149662104e+03 9.43e-04 3.15e-01 0.00e+00
#>   51 -3.3199915009186420e+03 9.26e-04 3.43e-01 0.00e+00
#>   52 -3.3199012567355298e+03 9.02e-04 3.70e-01 0.00e+00
#>   53 -3.3198152453866696e+03 8.72e-04 3.97e-01 0.00e+00
#>   54 -3.3197337427753932e+03 8.36e-04 4.23e-01 0.00e+00
#>   55 -3.3196569611795512e+03 7.96e-04 4.48e-01 0.00e+00
#>   56 -3.3195850568185319e+03 7.53e-04 4.70e-01 0.00e+00
#>   57 -3.3195181273360740e+03 7.07e-04 4.90e-01 0.00e+00
#>   58 -3.3194562051376988e+03 6.61e-04 5.06e-01 0.00e+00
#>   59 -3.3193992511845363e+03 6.13e-04 5.18e-01 0.00e+00
#>   60 -3.3193471521622532e+03 5.66e-04 5.26e-01 0.00e+00
#>   61 -3.3192997224172850e+03 5.20e-04 5.30e-01 0.00e+00
#>   62 -3.3192567108306430e+03 5.27e-04 5.29e-01 0.00e+00
#>   63 -3.3192178119040900e+03 5.60e-04 5.24e-01 0.00e+00
#>   64 -3.3191826797634758e+03 5.92e-04 5.14e-01 0.00e+00
#>   65 -3.3191509435299586e+03 6.22e-04 5.00e-01 0.00e+00
#>   66 -3.3191222225397241e+03 6.50e-04 4.83e-01 0.00e+00
#>   67 -3.3190961401428276e+03 6.77e-04 4.62e-01 0.00e+00
#>   68 -3.3190723351939041e+03 7.02e-04 4.40e-01 0.00e+00
#>   69 -3.3190504707686810e+03 7.25e-04 4.15e-01 0.00e+00
#>   70 -3.3190306208626421e+03 7.46e-04 3.91e-01 0.00e+00
#>   71 -3.3190123287869201e+03 7.65e-04 3.61e-01 0.00e+00
#>   72 -3.3189952323726811e+03 7.83e-04 3.29e-01 0.00e+00
#>   73 -3.3189790814910320e+03 7.98e-04 2.98e-01 0.00e+00
#>   74 -3.3189636722150353e+03 8.11e-04 2.68e-01 0.00e+00
#>   75 -3.3189495803885711e+03 8.23e-04 2.42e-01 0.00e+00
#>   76 -3.3189369045830822e+03 8.26e-04 2.17e-01 0.00e+00
#>   77 -3.3189249513249533e+03 8.23e-04 1.95e-01 0.00e+00
#>   78 -3.3189136101729437e+03 8.21e-04 1.75e-01 0.00e+00
#>   79 -3.3189027986816673e+03 8.19e-04 1.58e-01 0.00e+00
#>   80 -3.3188924524533950e+03 8.16e-04 1.41e-01 0.00e+00
#>   81 -3.3188825206340175e+03 8.11e-04 1.27e-01 0.00e+00
#>   82 -3.3188729628058150e+03 8.06e-04 1.14e-01 0.00e+00
#>   83 -3.3188637466295122e+03 7.99e-04 1.02e-01 0.00e+00
#>   84 -3.3188548460179109e+03 7.92e-04 9.21e-02 0.00e+00
#>   85 -3.3188462397175522e+03 7.84e-04 8.30e-02 0.00e+00
#>   86 -3.3188379102090967e+03 7.76e-04 7.50e-02 0.00e+00
#>   87 -3.3188298428569324e+03 7.67e-04 6.79e-02 0.00e+00
#>   88 -3.3188220252530723e+03 7.57e-04 6.17e-02 0.00e+00
#>   89 -3.3188144467119255e+03 7.47e-04 5.61e-02 0.00e+00
#>   90 -3.3188070978817818e+03 7.37e-04 5.13e-02 0.00e+00
#>   91 -3.3187999704463386e+03 7.26e-04 4.70e-02 0.00e+00
#>   92 -3.3187930568955353e+03 7.16e-04 4.32e-02 0.00e+00
#>   93 -3.3187863503496746e+03 7.05e-04 3.98e-02 0.00e+00
#>   94 -3.3187798444244745e+03 6.94e-04 3.68e-02 0.00e+00
#>   95 -3.3187735331275790e+03 6.83e-04 3.41e-02 0.00e+00
#>   96 -3.3187674107792755e+03 6.72e-04 3.17e-02 0.00e+00
#>   97 -3.3187614719518469e+03 6.60e-04 2.96e-02 0.00e+00
#>   98 -3.3187557114232904e+03 6.49e-04 2.77e-02 0.00e+00
#>   99 -3.3187501241421583e+03 6.38e-04 2.59e-02 0.00e+00
#>  100 -3.3187447052009752e+03 6.27e-04 2.44e-02 0.00e+00