Initialize Ultimate Deconvolution Model

Initialize an Ultimate Deconvolution model. See ud_fit for background and model definition.

ud_init(
  X,
  V = diag(ncol(X)),
  n_rank1,
  n_unconstrained,
  U_scaled = list(indep = diag(ncol(X)), iden = matrix(1, ncol(X), ncol(X))),
  U_rank1,
  U_unconstrained,
  control = list()
)

Arguments

X
The n x m data matrix, in which each row of the matrix is an m-dimensional data point. The number of rows and columns should be 2 or more.

V
Either an m x m matrix giving the initial estimate of the residual covariance matrix, or a list of m x m "standard error" matrices, one for each data point.

n_rank1
A non-negative integer specifying the number of rank-1 covariance matrices included in the mixture prior. Initial estimates of the m x m rank-1 covariance matrices are generated at random. At most one of n_rank1 and U_rank1 should be provided. If neither are specified, 4 rank-1 matrices will be used.

n_unconstrained
A non-negative integer specifying the number of unconstrained covariance matrices included in the mixture prior. Initial estimates of the m x m covariance matrices are generated at random. At most one of n_unconstrained and U_unconstrained should be provided. If neither are specified, 4 unconstrained matrices are used.

U_scaled
A list specifying initial estimates of the scaled covariance matrices in the mixture prior.

U_rank1
A list specifying initial estimates of the rank-1 matrices in the mixture prior. At most one of n_rank1 and U_rank1 should be provided. If U_rank1 is not given, the rank-1 covariates are initialized at random.

U_unconstrained
A list specifying initial estimates of the unconstrained matrices in the mixture prior. At most one of n_unconstrained and U_unconstrained should be provided. If U_unconstrained is not given, the matrices are initialized at random.

control
A list of parameters controlling the behaviour of the model initialization. See ud_fit for details.

Value

An object of class "ud_fit". See ud_fit for details.